refactor(api): consolidate auth tests into a single /health/auth probe
Tests now exercise the auth flow through HTTP requests and assert on
response bodies only - no DB or signal-state assertions.

- Add /health/auth probe route that returns the resolved AuthContext.
  The route is configured with acceptAnySandboxCapability so a single
  endpoint covers session, PAT, sandbox, and zero credentials.
- Replace the previous apiHealthAuth handler with the probe; drop the
  now-unused apiHealthAuth export and the healthAuthContract route
  registration.
- Replace auth-context.test.ts and auth-route.test.ts with a single
  health-auth-probe.test.ts that hits /health/auth with each token
  type, asserts the resolved context shape, and checks 401 paths
  (missing credentials, invalid token, missing org membership).
- Drop unused createAuthContext and apiKeyAuthContext exports from
  auth-context.ts; the only required entry point is now
  requiredAuthContext (used by authRoute).
