+++ turbo/apps/api/src/signals/routes/health.ts
@@ -1,19 +1,7 @@
 import { computed } from "ccstate";
-import {
-  type HealthAuthRouteResponse,
-  type HealthRouteResponse,
-} from "@vm0/api-contracts/contracts";
-
-import { authRoute } from "../auth/auth-route";
+import { type HealthRouteResponse } from "@vm0/api-contracts/contracts";
 
 export const apiHealth$ = computed<Promise<HealthRouteResponse>>(async () => {
   await Promise.resolve();
   return { status: 200, body: { status: "ok" } };
 });
-
-export const apiHealthAuth$ = authRoute(
-  {},
-  computed((): HealthAuthRouteResponse => {
-    return { status: 200 as const, body: { status: "ok" } };
-  }),
-);
